pyproject.toml
[project]
name = "sartor-ad-engine"
version = "0.1.0"
description = "Automated multi-agent system for generating personalized static ads for eCommerce products"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Image processing
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    
    # Configuration
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "agents*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
